+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add support for specifying different program entrypoints via variable `lsp-dart-program-entrypoints` used by DAP for example.
 * Drop support for Eamcs 26.1 and 26.2
+* Improve lsp-dart-flutter-outline performance a lot, requesting code actions lazily.
 
 ## 1.20.0